burn: all the old code
we are going stream first so we dont need any of that bidirectional
stuff
