+++ lib/llm_gateway/adapters/anthropic/input_mapper.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "bidirectional_message_mapper"
-
 module LlmGateway
   module Adapters
     module Anthropic
@@ -14,42 +12,123 @@ module LlmGateway
           }
         end
 
-        private
+        def self.map_content(content)
+          content = { type: "text", text: content } unless content.is_a?(Hash)
 
-        def self.map_messages(messages)
-          return messages unless messages
-
-          message_mapper = BidirectionalMessageMapper.new(LlmGateway::DIRECTION_IN)
-
-          messages.map do |msg|
-            msg = msg.merge(role: "user") if msg[:role] == "developer"
-
-            content = if msg[:content].is_a?(Array)
-                msg[:content].map do |content|
-                  message_mapper.map_content(content)
-                end
-            else
-              [ message_mapper.map_content(msg[:content]) ]
-            end
-
-            {
-              role: msg[:role],
-              content: content
-            }
+          case content[:type]
+          when "text"
+            map_text_content(content)
+          when "file"
+            map_file_content(content)
+          when "image"
+            map_image_content(content)
+          when "tool_use"
+            map_tool_use_content(content)
+          when "tool_result"
+            map_tool_result_content(content)
+          when "thinking", "reasoning"
+            map_reasoning_content(content)
+          else
+            content
           end
         end
 
-        def self.map_system(system)
-          if !system || system.empty?
-            nil
-          elsif system.length == 1 && system.first[:role] == "system"
-            # If we have a single system message, convert to Claude format
-            mapped = { type: "text", text: system.first[:content] }
-            mapped[:cache_control] = system.first[:cache_control] if system.first[:cache_control]
-            [ mapped ]
-          else
-            # For multiple messages or non-standard format, pass through
-            system
+        class << self
+          private
+
+          def map_messages(messages)
+            return messages unless messages
+
+            messages.map do |msg|
+              msg = msg.merge(role: "user") if msg[:role] == "developer"
+
+              content = if msg[:content].is_a?(Array)
+                msg[:content].map { |content| map_content(content) }
+              else
+                [ map_content(msg[:content]) ]
+              end
+
+              {
+                role: msg[:role],
+                content: content
+              }
+            end
+          end
+
+          def map_system(system)
+            if !system || system.empty?
+              nil
+            elsif system.length == 1 && system.first[:role] == "system"
+              mapped = { type: "text", text: system.first[:content] }
+              mapped[:cache_control] = system.first[:cache_control] if system.first[:cache_control]
+              [ mapped ]
+            else
+              system
+            end
+          end
+
+          def map_text_content(content)
+            result = {
+              type: "text",
+              text: content[:text]
+            }
+            result[:cache_control] = content[:cache_control] if content[:cache_control]
+            result
+          end
+
+          def map_file_content(content)
+            {
+              type: "document",
+              source: {
+                data: content[:data],
+                type: "text",
+                media_type: content[:media_type]
+              }
+            }
+          end
+
+          def map_image_content(content)
+            {
+              type: "image",
+              source: {
+                data: content[:data],
+                type: "base64",
+                media_type: content[:media_type]
+              }
+            }
+          end
+
+          def map_tool_use_content(content)
+            {
+              type: "tool_use",
+              id: content[:id],
+              name: content[:name],
+              input: content[:input]
+            }
+          end
+
+          def map_tool_result_content(content)
+            mapped_content = content[:content]
+            if mapped_content.is_a?(Array)
+              mapped_content = mapped_content.map do |item|
+                item.is_a?(Hash) ? map_content(item.transform_keys(&:to_sym)) : item
+              end
+            end
+
+            {
+              type: "tool_result",
+              tool_use_id: content[:tool_use_id],
+              content: mapped_content
+            }
+          end
+
+          def map_reasoning_content(content)
+            result = {
+              type: "thinking",
+              thinking: content[:reasoning]
+            }
+            result[:signature] = content[:signature] unless content[:signature].nil?
+            result
           end
         end
       end
