+++ lib/llm_gateway/adapters/openai/chat_completions_adapter.rb
@@ -4,7 +4,6 @@ require_relative "../adapter"
 require_relative "acts_like_chat_completions"
 require_relative "chat_completions/input_mapper"
 require_relative "chat_completions/input_message_sanitizer"
-require_relative "chat_completions/output_mapper"
 require_relative "chat_completions/option_mapper"
 require_relative "file_output_mapper"
 require_relative "chat_completions/stream_mapper"
