+++ lib/llm_gateway/adapters/openai_codex/responses_adapter.rb
@@ -2,7 +2,6 @@
 
 require_relative "../adapter"
 require_relative "../openai/acts_like_responses"
-require_relative "../openai/responses/output_mapper"
 require_relative "option_mapper"
 require_relative "../openai/responses/stream_mapper"
 require_relative "../openai/file_output_mapper"
@@ -23,10 +22,6 @@ module LlmGateway
 
         def option_mapper
           OptionMapper
-        end
-
-        def perform_chat(messages, tools:, system:, **options)
-          client.chat_codex(messages, tools: tools, system: system, **options)
         end
 
         def perform_stream(messages, tools:, system:, **options, &block)
