--- test/fixtures/vcr_cassettes/integration/document_usage_test_rb/test_claude_document_usage.yml
@@ -1,86 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://api.anthropic.com/v1/messages
-    body:
-      encoding: UTF-8
-      string: '{"model":"claude-sonnet-4-20250514","max_tokens":4096,"messages":[{"role":"user","content":[{"type":"text","text":"return
-        the content of the document exactly"},{"type":"document","source":{"data":"abc\n","type":"text","media_type":"text/plain"}}]}]}'
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-      Host:
-      - api.anthropic.com
-      Anthropic-Version:
-      - '2023-06-01'
-      Content-Type:
-      - application/json
-      X-Api-Key:
-      - "<BEARER_TOKEN>"
-      Anthropic-Beta:
-      - code-execution-2025-05-22,files-api-2025-04-14
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Sat, 14 Feb 2026 14:43:33 GMT
-      Content-Type:
-      - application/json
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Anthropic-Ratelimit-Input-Tokens-Limit:
-      - '450000'
-      Anthropic-Ratelimit-Input-Tokens-Remaining:
-      - '450000'
-      Anthropic-Ratelimit-Input-Tokens-Reset:
-      - '2026-02-14T14:43:33Z'
-      Anthropic-Ratelimit-Output-Tokens-Limit:
-      - '90000'
-      Anthropic-Ratelimit-Output-Tokens-Remaining:
-      - '90000'
-      Anthropic-Ratelimit-Output-Tokens-Reset:
-      - '2026-02-14T14:43:33Z'
-      Anthropic-Ratelimit-Requests-Limit:
-      - '1000'
-      Anthropic-Ratelimit-Requests-Remaining:
-      - '999'
-      Anthropic-Ratelimit-Requests-Reset:
-      - '2026-02-14T14:43:32Z'
-      Anthropic-Ratelimit-Tokens-Limit:
-      - '540000'
-      Anthropic-Ratelimit-Tokens-Remaining:
-      - '540000'
-      Anthropic-Ratelimit-Tokens-Reset:
-      - '2026-02-14T14:43:33Z'
-      Request-Id:
-      - req_011CY892ETh5vYPWe3H7mQCs
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      Anthropic-Organization-Id:
-      - 0fcc9bdc-70ee-4701-a64f-cdc9e356dfa8
-      Server:
-      - cloudflare
-      X-Envoy-Upstream-Service-Time:
-      - '1065'
-      Cf-Cache-Status:
-      - DYNAMIC
-      X-Robots-Tag:
-      - none
-      Content-Security-Policy:
-      - default-src 'none'; frame-ancestors 'none'
-      Cf-Ray:
-      - 9cdd56a0ecdc5545-SIN
-    body:
-      encoding: ASCII-8BIT
-      string: '{"model":"claude-sonnet-4-20250514","id":"msg_01JRufPbQe35r7M1e5ndx3Ht","type":"message","role":"assistant","content":[{"type":"text","text":"abc"}],"stop_reason":"end_turn","stop_sequence":null,"usage":{"input_tokens":53,"cache_creation_input_tokens":0,"cache_read_input_tokens":0,"cache_creation":{"ephemeral_5m_input_tokens":0,"ephemeral_1h_input_tokens":0},"output_tokens":4,"service_tier":"standard","inference_geo":"not_available"}}'
-  recorded_at: Sat, 14 Feb 2026 14:43:33 GMT
-recorded_with: VCR 6.4.0
