--- test/fixtures/vcr_cassettes/integration/document_usage_test_rb/test_cluade_upload_file.yml
@@ -1,61 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://api.anthropic.com/v1/files
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-      Host:
-      - api.anthropic.com
-      Content-Type:
-      - multipart/form-data
-      Anthropic-Version:
-      - '2023-06-01'
-      X-Api-Key:
-      - "<BEARER_TOKEN>"
-      Anthropic-Beta:
-      - code-execution-2025-05-22,files-api-2025-04-14
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Sat, 14 Feb 2026 14:43:32 GMT
-      Content-Type:
-      - application/json
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Request-Id:
-      - req_011CY8929T4hF4hGpUAGZfpk
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      Anthropic-Organization-Id:
-      - 0fcc9bdc-70ee-4701-a64f-cdc9e356dfa8
-      Server:
-      - cloudflare
-      X-Envoy-Upstream-Service-Time:
-      - '415'
-      Cf-Cache-Status:
-      - DYNAMIC
-      X-Robots-Tag:
-      - none
-      Content-Security-Policy:
-      - default-src 'none'; frame-ancestors 'none'
-      Cf-Ray:
-      - 9cdd5699995eafdd-SIN
-    body:
-      encoding: ASCII-8BIT
-      string: '{"type":"file","id":"file_011CY8929ssFxi4Uod8urMEF","size_bytes":13,"created_at":"2026-02-14T14:43:31.666000Z","filename":"test.txt","mime_type":"application/octet-stream","downloadable":false}'
-  recorded_at: Sat, 14 Feb 2026 14:43:32 GMT
-recorded_with: VCR 6.4.0
