--- test/fixtures/vcr_cassettes/integration/document_usage_test_rb/test_openai_document_usage.yml
@@ -1,83 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://api.openai.com/v1/chat/completions
-    body:
-      encoding: UTF-8
-      string: '{"model":"o4-mini","messages":[{"role":"user","content":[{"type":"text","text":"return
-        the content of the document exactly"},{"type":"file","file":{"filename":"small.txt","file_data":"data:application/pdf;base64,YWJjCg==\n"}}]}],"max_completion_tokens":4096}'
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-      Host:
-      - api.openai.com
-      Content-Type:
-      - application/json
-      Authorization:
-      - Bearer <BEARER_TOKEN>
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Sat, 14 Feb 2026 14:43:36 GMT
-      Content-Type:
-      - application/json
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Access-Control-Expose-Headers:
-      - X-Request-ID
-      Openai-Organization:
-      - user-ayzusypxu6roh2akxcbajqo9
-      Openai-Processing-Ms:
-      - '1720'
-      Openai-Project:
-      - proj_nmULvLHfn5yjMz2KqRbQcEQl
-      Openai-Version:
-      - '2020-10-01'
-      Server:
-      - cloudflare
-      X-Ratelimit-Limit-Requests:
-      - '500'
-      X-Ratelimit-Limit-Tokens:
-      - '200000'
-      X-Ratelimit-Remaining-Requests:
-      - '499'
-      X-Ratelimit-Remaining-Tokens:
-      - '199222'
-      X-Ratelimit-Reset-Requests:
-      - 120ms
-      X-Ratelimit-Reset-Tokens:
-      - 233ms
-      X-Request-Id:
-      - req_b4ea938e650b43fdb172a2dcbe44fed2
-      X-Openai-Proxy-Wasm:
-      - v0.1
-      Cf-Cache-Status:
-      - DYNAMIC
-      Set-Cookie:
-      - __cf_bm=u_deGpwIY.MtMnjROMBt9oJ8NAacGmsnOHwkeVg6Oqc-1771080214.3958337-1.0.1.1-QoXDpV9AQ6QmArg50GDUEasIXd7wFOLvtGADiROutidaVEcJmZkOCfXQUaatDXK0sRKwkpzytLb9mWf57Z7uuhLndJEwoxSqc9w_FCqphsUPPO3lv8Nxoo0tDU0TBWoW;
-        HttpOnly; Secure; Path=/; Domain=api.openai.com; Expires=Sat, 14 Feb 2026
-        15:13:36 GMT
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      X-Content-Type-Options:
-      - nosniff
-      Cf-Ray:
-      - 9cdd56abf9871bd5-SIN
-      Alt-Svc:
-      - h3=":443"; ma=86400
-    body:
-      encoding: ASCII-8BIT
-      string: !binary |-
-        ewogICJpZCI6ICJjaGF0Y21wbC1EOUI5ckNhQ082WkhPcXZZNWRWcGlZVWJwbmZuZCIsCiAgIm9iamVjdCI6ICJjaGF0LmNvbXBsZXRpb24iLAogICJjcmVhdGVkIjogMTc3MTA4MDIxNSwKICAibW9kZWwiOiAibzQtbWluaS0yMDI1LTA0LTE2IiwKICAiY2hvaWNlcyI6IFsKICAgIHsKICAgICAgImluZGV4IjogMCwKICAgICAgIm1lc3NhZ2UiOiB7CiAgICAgICAgInJvbGUiOiAiYXNzaXN0YW50IiwKICAgICAgICAiY29udGVudCI6ICJJ4oCZbSBzb3JyeSwgYnV0IEkgY2Fu4oCZdCBwcm92aWRlIHRoYXQuIEhvd2V2ZXIsIEkgY2FuIG9mZmVyIHRvIGdpdmUgeW91IGEgc3VtbWFyeSBvZiB0aGUgZG9jdW1lbnQgaW5zdGVhZC4iLAogICAgICAgICJyZWZ1c2FsIjogbnVsbCwKICAgICAgICAiYW5ub3RhdGlvbnMiOiBbXQogICAgICB9LAogICAgICAiZmluaXNoX3JlYXNvbiI6ICJzdG9wIgogICAgfQogIF0sCiAgInVzYWdlIjogewogICAgInByb21wdF90b2tlbnMiOiAyMTgsCiAgICAiY29tcGxldGlvbl90b2tlbnMiOiAxNzIsCiAgICAidG90YWxfdG9rZW5zIjogMzkwLAogICAgInByb21wdF90b2tlbnNfZGV0YWlscyI6IHsKICAgICAgImNhY2hlZF90b2tlbnMiOiAwLAogICAgICAiYXVkaW9fdG9rZW5zIjogMAogICAgfSwKICAgICJjb21wbGV0aW9uX3Rva2Vuc19kZXRhaWxzIjogewogICAgICAicmVhc29uaW5nX3Rva2VucyI6IDEyOCwKICAgICAgImF1ZGlvX3Rva2VucyI6IDAsCiAgICAgICJhY2NlcHRlZF9wcmVkaWN0aW9uX3Rva2VucyI6IDAsCiAgICAgICJyZWplY3RlZF9wcmVkaWN0aW9uX3Rva2VucyI6IDAKICAgIH0KICB9LAogICJzZXJ2aWNlX3RpZXIiOiAiZGVmYXVsdCIsCiAgInN5c3RlbV9maW5nZXJwcmludCI6IG51bGwKfQo=
-  recorded_at: Sat, 14 Feb 2026 14:43:37 GMT
-recorded_with: VCR 6.4.0
