--- test/fixtures/vcr_cassettes/integration/document_usage_test_rb/test_openai_upload_file.yml
@@ -1,80 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://api.openai.com/v1/files
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-      Host:
-      - api.openai.com
-      Content-Type:
-      - multipart/form-data
-      Authorization:
-      - Bearer <BEARER_TOKEN>
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Sat, 14 Feb 2026 14:43:38 GMT
-      Content-Type:
-      - application/json
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Server:
-      - cloudflare
-      X-Request-Id:
-      - req_ca0f7b73c0a74bc7870c4f50afc54844
-      Openai-Processing-Ms:
-      - '395'
-      Openai-Version:
-      - '2020-10-01'
-      Openai-Organization:
-      - user-ayzusypxu6roh2akxcbajqo9
-      Openai-Project:
-      - proj_nmULvLHfn5yjMz2KqRbQcEQl
-      Access-Control-Allow-Origin:
-      - "*"
-      X-Openai-Proxy-Wasm:
-      - v0.1
-      Cf-Cache-Status:
-      - DYNAMIC
-      Set-Cookie:
-      - __cf_bm=cOtqZHFoozOMb8pXJx9vTsTfQSb3ZihwQONAoV4pB8U-1771080217.5568936-1.0.1.1-YWYO2PQNq5D0Bm9GOkdaH.OQG.Iju2LrMxSAGAAxs1HZVLgHQui1Wfd70moSKkKW1ELkWFa0lV.hUfV3BJqxbA44ITw.HjauK8qg2ydzv8I89VQf5J5WvCFtpMGjAoGM;
-        HttpOnly; Secure; Path=/; Domain=api.openai.com; Expires=Sat, 14 Feb 2026
-        15:13:38 GMT
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      X-Content-Type-Options:
-      - nosniff
-      Cf-Ray:
-      - 9cdd56bfb89cf543-SIN
-      Alt-Svc:
-      - h3=":443"; ma=86400
-    body:
-      encoding: ASCII-8BIT
-      string: |
-        {
-          "object": "file",
-          "id": "file-CEdobHN6xZ5G2jdSDKdft4",
-          "purpose": "user_data",
-          "filename": "test.txt",
-          "bytes": 13,
-          "created_at": 1771080217,
-          "expires_at": null,
-          "status": "processed",
-          "status_details": null
-        }
-  recorded_at: Sat, 14 Feb 2026 14:43:38 GMT
-recorded_with: VCR 6.4.0
