--- test/fixtures/vcr_cassettes/integration/gateway_test_rb/test_claude_weather_with_pirate_system.yml
@@ -1,89 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://api.anthropic.com/v1/messages
-    body:
-      encoding: UTF-8
-      string: '{"model":"claude-sonnet-4-20250514","max_tokens":4096,"messages":[{"role":"user","content":[{"type":"text","text":"What''s
-        the weather in Singapore? reply in 10 words and no special characters"}]}],"tools":[{"name":"get_weather","description":"Get
-        current weather for a location","input_schema":{"type":"object","properties":{"location":{"type":"string","description":"City
-        name"}},"required":["location"]}}],"system":[{"type":"text","text":"Talk like
-        a pirate"}]}'
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-      Host:
-      - api.anthropic.com
-      Anthropic-Version:
-      - '2023-06-01'
-      Content-Type:
-      - application/json
-      X-Api-Key:
-      - "<BEARER_TOKEN>"
-      Anthropic-Beta:
-      - code-execution-2025-05-22,files-api-2025-04-14
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Sat, 14 Feb 2026 14:44:49 GMT
-      Content-Type:
-      - application/json
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Anthropic-Ratelimit-Input-Tokens-Limit:
-      - '450000'
-      Anthropic-Ratelimit-Input-Tokens-Remaining:
-      - '450000'
-      Anthropic-Ratelimit-Input-Tokens-Reset:
-      - '2026-02-14T14:44:48Z'
-      Anthropic-Ratelimit-Output-Tokens-Limit:
-      - '90000'
-      Anthropic-Ratelimit-Output-Tokens-Remaining:
-      - '90000'
-      Anthropic-Ratelimit-Output-Tokens-Reset:
-      - '2026-02-14T14:44:49Z'
-      Anthropic-Ratelimit-Requests-Limit:
-      - '1000'
-      Anthropic-Ratelimit-Requests-Remaining:
-      - '999'
-      Anthropic-Ratelimit-Requests-Reset:
-      - '2026-02-14T14:44:47Z'
-      Anthropic-Ratelimit-Tokens-Limit:
-      - '540000'
-      Anthropic-Ratelimit-Tokens-Remaining:
-      - '540000'
-      Anthropic-Ratelimit-Tokens-Reset:
-      - '2026-02-14T14:44:48Z'
-      Request-Id:
-      - req_011CY897kGTd6EvM3eGo66nS
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      Anthropic-Organization-Id:
-      - 0fcc9bdc-70ee-4701-a64f-cdc9e356dfa8
-      Server:
-      - cloudflare
-      X-Envoy-Upstream-Service-Time:
-      - '2215'
-      Cf-Cache-Status:
-      - DYNAMIC
-      X-Robots-Tag:
-      - none
-      Content-Security-Policy:
-      - default-src 'none'; frame-ancestors 'none'
-      Cf-Ray:
-      - 9cdd58746a75aa1c-SIN
-    body:
-      encoding: ASCII-8BIT
-      string: '{"model":"claude-sonnet-4-20250514","id":"msg_01CW95fcbDQSzSZPWus2tzgv","type":"message","role":"assistant","content":[{"type":"tool_use","id":"toolu_01RmbpBz25gb9wziuBZjE8re","name":"get_weather","input":{"location":"Singapore"},"caller":{"type":"direct"}}],"stop_reason":"tool_use","stop_sequence":null,"usage":{"input_tokens":404,"cache_creation_input_tokens":0,"cache_read_input_tokens":0,"cache_creation":{"ephemeral_5m_input_tokens":0,"ephemeral_1h_input_tokens":0},"output_tokens":53,"service_tier":"standard","inference_geo":"not_available"}}'
-  recorded_at: Sat, 14 Feb 2026 14:44:49 GMT
-recorded_with: VCR 6.4.0
