--- test/fixtures/vcr_cassettes/integration/gateway_test_rb/test_groq_simple_message_without_tools.yml
@@ -1,80 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://api.groq.com/openai/v1/chat/completions
-    body:
-      encoding: UTF-8
-      string: '{"model":"llama-3.3-70b-versatile","messages":[{"role":"system","content":"Talk
-        like a pirate"},{"role":"user","content":[{"type":"text","text":"What''s the
-        weather in Singapore? reply in 10 words and no special characters"}]}],"temperature":0,"max_completion_tokens":4096,"response_format":{"type":"text"},"tools":null}'
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-      Host:
-      - api.groq.com
-      Content-Type:
-      - application/json
-      Authorization:
-      - Bearer <BEARER_TOKEN>
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Sat, 14 Feb 2026 14:44:50 GMT
-      Content-Type:
-      - application/json
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Cache-Control:
-      - private, max-age=0, no-store, no-cache, must-revalidate
-      Server:
-      - cloudflare
-      Vary:
-      - Origin
-      X-Groq-Region:
-      - bom
-      X-Ratelimit-Limit-Requests:
-      - '500000'
-      X-Ratelimit-Limit-Tokens:
-      - '300000'
-      X-Ratelimit-Remaining-Requests:
-      - '499999'
-      X-Ratelimit-Remaining-Tokens:
-      - '299944'
-      X-Ratelimit-Reset-Requests:
-      - 172ms
-      X-Ratelimit-Reset-Tokens:
-      - 11ms
-      X-Request-Id:
-      - req_01khe9s41te8mvvqhzsskfpr0n
-      Set-Cookie:
-      - __cf_bm=sUtkz0uFDox2KwwuLTSlItJeCWKxvzJpD1WJWskmqZ4-1771080290.3264537-1.0.1.1-gC9K1TNc5lF4vyXu24hMusGte0j43rOOlVzFm0YXv0R2NG4GgvMoX4T7VPm7.JUl5VHwI.uecuPwnt52tY_wuv83fZ3HeohqCZEgbgaDEWWq8vzbRUDLqK_EiczkOmYc;
-        HttpOnly; Secure; Path=/; Domain=groq.com; Expires=Sat, 14 Feb 2026 15:14:50
-        GMT
-      Via:
-      - 1.1 google
-      Strict-Transport-Security:
-      - max-age=15552000
-      Cf-Cache-Status:
-      - DYNAMIC
-      Cf-Ray:
-      - 9cdd58868b199cb0-SIN
-      Alt-Svc:
-      - h3=":443"; ma=86400
-    body:
-      encoding: ASCII-8BIT
-      string: '{"id":"chatcmpl-439771d5-1bd8-4bec-a2ca-94dd2a39d671","object":"chat.completion","created":1771080290,"model":"llama-3.3-70b-versatile","choices":[{"index":0,"message":{"role":"assistant","content":"The
-        weather in Singapore is usually hot and humid"},"logprobs":null,"finish_reason":"stop"}],"usage":{"queue_time":0.050896314,"prompt_tokens":55,"prompt_time":0.004719656,"completion_tokens":10,"completion_time":0.029106236,"total_tokens":65,"total_time":0.033825892},"usage_breakdown":null,"system_fingerprint":"fp_dae98b5ecb","x_groq":{"id":"req_01khe9s41te8mvvqhzsskfpr0n","seed":1517367084},"service_tier":"on_demand"}
-
-        '
-  recorded_at: Sat, 14 Feb 2026 14:44:50 GMT
-recorded_with: VCR 6.4.0
