--- test/fixtures/vcr_cassettes/integration/gateway_test_rb/test_groq_weather_with_pirate_system.yml
@@ -1,82 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://api.groq.com/openai/v1/chat/completions
-    body:
-      encoding: UTF-8
-      string: '{"model":"llama-3.3-70b-versatile","messages":[{"role":"system","content":"Talk
-        like a pirate"},{"role":"user","content":[{"type":"text","text":"What''s the
-        weather in Singapore? reply in 10 words and no special characters"}]}],"temperature":0,"max_completion_tokens":4096,"response_format":{"type":"text"},"tools":[{"type":"function","function":{"name":"get_weather","description":"Get
-        current weather for a location","parameters":{"type":"object","properties":{"location":{"type":"string","description":"City
-        name"}},"required":["location"]}}}]}'
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-      Host:
-      - api.groq.com
-      Content-Type:
-      - application/json
-      Authorization:
-      - Bearer <BEARER_TOKEN>
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Sat, 14 Feb 2026 14:44:59 GMT
-      Content-Type:
-      - application/json
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Cache-Control:
-      - private, max-age=0, no-store, no-cache, must-revalidate
-      Server:
-      - cloudflare
-      Vary:
-      - Origin
-      X-Groq-Region:
-      - bom
-      X-Ratelimit-Limit-Requests:
-      - '500000'
-      X-Ratelimit-Limit-Tokens:
-      - '300000'
-      X-Ratelimit-Remaining-Requests:
-      - '499999'
-      X-Ratelimit-Remaining-Tokens:
-      - '299762'
-      X-Ratelimit-Reset-Requests:
-      - 172ms
-      X-Ratelimit-Reset-Tokens:
-      - 47ms
-      X-Request-Id:
-      - req_01khe9sd23ftg8bfxgyf4jecsf
-      Set-Cookie:
-      - __cf_bm=VFRzy0IfZcRSMlGJ0HHkQE7xAKrU0E0kDGQxcDtvwFY-1771080299.5442393-1.0.1.1-P7POjGGuMIqj068Nq8Pa8AvOZZNUv2gCg0RNFjJ4Cozg4x3wie5mKWHoOdNzKY5cDYywuOsqndX.vnl5YOAgAWjkW_8HXc_0M.ZuuDhLMPxAsBAy8XDbk7c7d2VP5D2Y;
-        HttpOnly; Secure; Path=/; Domain=groq.com; Expires=Sat, 14 Feb 2026 15:14:59
-        GMT
-      Via:
-      - 1.1 google
-      Strict-Transport-Security:
-      - max-age=15552000
-      Cf-Cache-Status:
-      - DYNAMIC
-      Cf-Ray:
-      - 9cdd58c01a58fd2a-SIN
-      Alt-Svc:
-      - h3=":443"; ma=86400
-    body:
-      encoding: ASCII-8BIT
-      string: '{"id":"chatcmpl-1b8ec5a5-291e-4516-a315-81aad31d94ae","object":"chat.completion","created":1771080299,"model":"llama-3.3-70b-versatile","choices":[{"index":0,"message":{"role":"assistant","content":"Get
-        the weather in Singapore right now please matey"},"logprobs":null,"finish_reason":"stop"}],"usage":{"queue_time":0.048110075,"prompt_tokens":237,"prompt_time":0.023137605,"completion_tokens":11,"completion_time":0.050607458,"total_tokens":248,"total_time":0.073745063},"usage_breakdown":null,"system_fingerprint":"fp_c06d5113ec","x_groq":{"id":"req_01khe9sd23ftg8bfxgyf4jecsf","seed":539579587},"service_tier":"on_demand"}
-
-        '
-  recorded_at: Sat, 14 Feb 2026 14:44:59 GMT
-recorded_with: VCR 6.4.0
