--- test/fixtures/vcr_cassettes/integration/gateway_test_rb/test_groq_weather_with_pirate_system_with_tool_usage.yml
@@ -1,82 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://api.groq.com/openai/v1/chat/completions
-    body:
-      encoding: UTF-8
-      string: '{"model":"llama-3.3-70b-versatile","messages":[{"role":"system","content":"Talk
-        like a pirate"},{"role":"user","content":[{"type":"text","text":"What''s the
-        weather in Singapore? reply in 10 words and no special characters"}]}],"temperature":0,"max_completion_tokens":4096,"response_format":{"type":"text"},"tools":[{"type":"function","function":{"name":"get_weather","description":"Get
-        current weather for a location","parameters":{"type":"object","properties":{"location":{"type":"string","description":"City
-        name"}},"required":["location"]}}}]}'
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-      Host:
-      - api.groq.com
-      Content-Type:
-      - application/json
-      Authorization:
-      - Bearer <BEARER_TOKEN>
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Sat, 14 Feb 2026 14:44:39 GMT
-      Content-Type:
-      - application/json
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Cache-Control:
-      - private, max-age=0, no-store, no-cache, must-revalidate
-      Server:
-      - cloudflare
-      Vary:
-      - Origin
-      X-Groq-Region:
-      - bom
-      X-Ratelimit-Limit-Requests:
-      - '500000'
-      X-Ratelimit-Limit-Tokens:
-      - '300000'
-      X-Ratelimit-Remaining-Requests:
-      - '499999'
-      X-Ratelimit-Remaining-Tokens:
-      - '299762'
-      X-Ratelimit-Reset-Requests:
-      - 172ms
-      X-Ratelimit-Reset-Tokens:
-      - 47ms
-      X-Request-Id:
-      - req_01khe9rsn1e8aaym6hr8p0aj0c
-      Via:
-      - 1.1 google
-      Cf-Cache-Status:
-      - DYNAMIC
-      Set-Cookie:
-      - __cf_bm=my4fOzn8RcPI9WxTMb7wwr2tN4RXZa13L7UrsVLcj94-1771080279.6735911-1.0.1.1-dTJaq7YSBgWV394vlrM.LQOhBMfVZn_aGJhwOvP2WRuiVuLgfl_PLQWURocX67I5GGfNe1rFKglLlrf2Q9sX_QwaueXJHf7YJLqaba5cItq7McIjONmjlg_YxiIY4isE;
-        HttpOnly; Secure; Path=/; Domain=groq.com; Expires=Sat, 14 Feb 2026 15:14:39
-        GMT
-      Strict-Transport-Security:
-      - max-age=15552000
-      Cf-Ray:
-      - 9cdd5843fc2aa125-SIN
-      Alt-Svc:
-      - h3=":443"; ma=86400
-    body:
-      encoding: ASCII-8BIT
-      string: '{"id":"chatcmpl-24feaf4d-6743-4a62-89d2-8df4fef9fc8d","object":"chat.completion","created":1771080279,"model":"llama-3.3-70b-versatile","choices":[{"index":0,"message":{"role":"assistant","content":"Get
-        the weather in Singapore right now please matey"},"logprobs":null,"finish_reason":"stop"}],"usage":{"queue_time":0.049580286,"prompt_tokens":237,"prompt_time":0.018155824,"completion_tokens":11,"completion_time":0.050747983,"total_tokens":248,"total_time":0.068903807},"usage_breakdown":null,"system_fingerprint":"fp_dae98b5ecb","x_groq":{"id":"req_01khe9rsn1e8aaym6hr8p0aj0c","seed":187989687},"service_tier":"on_demand"}
-
-        '
-  recorded_at: Sat, 14 Feb 2026 14:44:39 GMT
-recorded_with: VCR 6.4.0
