--- test/fixtures/vcr_cassettes/integration/gateway_test_rb/test_openai_simple_message_without_tools.yml
@@ -1,118 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://api.openai.com/v1/chat/completions
-    body:
-      encoding: UTF-8
-      string: '{"model":"o4-mini","messages":[{"role":"developer","content":"Talk
-        like a pirate"},{"role":"user","content":[{"type":"text","text":"What''s the
-        weather in Singapore? reply in 10 words and no special characters"}]}],"max_completion_tokens":4096}'
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-      Host:
-      - api.openai.com
-      Content-Type:
-      - application/json
-      Authorization:
-      - Bearer <BEARER_TOKEN>
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Sat, 14 Feb 2026 14:44:46 GMT
-      Content-Type:
-      - application/json
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Access-Control-Expose-Headers:
-      - X-Request-ID
-      Openai-Organization:
-      - user-ayzusypxu6roh2akxcbajqo9
-      Openai-Processing-Ms:
-      - '5891'
-      Openai-Project:
-      - proj_nmULvLHfn5yjMz2KqRbQcEQl
-      Openai-Version:
-      - '2020-10-01'
-      Server:
-      - cloudflare
-      X-Ratelimit-Limit-Requests:
-      - '500'
-      X-Ratelimit-Limit-Tokens:
-      - '200000'
-      X-Ratelimit-Remaining-Requests:
-      - '499'
-      X-Ratelimit-Remaining-Tokens:
-      - '199973'
-      X-Ratelimit-Reset-Requests:
-      - 120ms
-      X-Ratelimit-Reset-Tokens:
-      - 8ms
-      X-Request-Id:
-      - req_322c210862e74e398ead1a9f76d01132
-      X-Openai-Proxy-Wasm:
-      - v0.1
-      Cf-Cache-Status:
-      - DYNAMIC
-      Set-Cookie:
-      - __cf_bm=pwieAI4a_r_4bWpiMwaPZv.Fl90HOR2BlhqeN9ExjTs-1771080280.5178664-1.0.1.1-LTtamle90wqXmuza3Iw_n724QrLoWJHujJfaNqJTI0xkEi9aXd9zHTztfDEUj6rULJw4uJoMl828cIH9dUhKmW829j747DmgPqYrveWIA0BES.M0Cmw0sau8GJ5H7Lwp;
-        HttpOnly; Secure; Path=/; Domain=api.openai.com; Expires=Sat, 14 Feb 2026
-        15:14:46 GMT
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      X-Content-Type-Options:
-      - nosniff
-      Cf-Ray:
-      - 9cdd58493b1f0b5f-SIN
-      Alt-Svc:
-      - h3=":443"; ma=86400
-    body:
-      encoding: ASCII-8BIT
-      string: |
-        {
-          "id": "chatcmpl-D9BAuAAUy5VJzXTrYyzu5AGeZh4R8",
-          "object": "chat.completion",
-          "created": 1771080280,
-          "model": "o4-mini-2025-04-16",
-          "choices": [
-            {
-              "index": 0,
-              "message": {
-                "role": "assistant",
-                "content": "Arr me hearty it be sunny humid in Singapore today",
-                "refusal": null,
-                "annotations": []
-              },
-              "finish_reason": "stop"
-            }
-          ],
-          "usage": {
-            "prompt_tokens": 29,
-            "completion_tokens": 860,
-            "total_tokens": 889,
-            "prompt_tokens_details": {
-              "cached_tokens": 0,
-              "audio_tokens": 0
-            },
-            "completion_tokens_details": {
-              "reasoning_tokens": 832,
-              "audio_tokens": 0,
-              "accepted_prediction_tokens": 0,
-              "rejected_prediction_tokens": 0
-            }
-          },
-          "service_tier": "default",
-          "system_fingerprint": null
-        }
-  recorded_at: Sat, 14 Feb 2026 14:44:47 GMT
-recorded_with: VCR 6.4.0
