--- test/fixtures/vcr_cassettes/integration/gateway_test_responses_test_rb/test_openai_responses_simple_message_transcript.yml
@@ -1,156 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://api.openai.com/v1/responses
-    body:
-      encoding: UTF-8
-      string: '{"model":"o4-mini","max_output_tokens":4096,"input":[{"role":"assistant","content":[{"type":"output_text","text":"Ahoy
-        matey Singapore be hot and humid with tropical showers"}]},{"role":"user","content":[{"type":"input_text","text":"what
-        did you think about during your last response"}]}],"instructions":"Talk like
-        a pirate"}'
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-      Host:
-      - api.openai.com
-      Content-Type:
-      - application/json
-      Authorization:
-      - Bearer <BEARER_TOKEN>
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Wed, 25 Mar 2026 14:40:35 GMT
-      Content-Type:
-      - application/json
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Server:
-      - cloudflare
-      X-Ratelimit-Limit-Requests:
-      - '500'
-      X-Ratelimit-Limit-Tokens:
-      - '200000'
-      X-Ratelimit-Remaining-Requests:
-      - '499'
-      X-Ratelimit-Remaining-Tokens:
-      - '199752'
-      X-Ratelimit-Reset-Requests:
-      - 120ms
-      X-Ratelimit-Reset-Tokens:
-      - 74ms
-      Openai-Version:
-      - '2020-10-01'
-      Openai-Organization:
-      - user-ayzusypxu6roh2akxcbajqo9
-      Openai-Project:
-      - proj_nmULvLHfn5yjMz2KqRbQcEQl
-      X-Request-Id:
-      - req_cb428232ce2c4683905fa5011d75bdfe
-      Openai-Processing-Ms:
-      - '4939'
-      Cf-Cache-Status:
-      - DYNAMIC
-      Set-Cookie:
-      - __cf_bm=ABshYdU0bQTzL2X4BN6iMZRWk917ynfX.pO9QFRugLU-1774449629.8530054-1.0.1.1-m2jkBBcaSKMjMtSfzmXapcaSdMUyFqP.a3sk89qqn5Ad99dsDq1p5p2CZTM6FaKEBuoJi_5dvN7QfyLlo9AxM_by5aNTfaCtHqWMCEtT9X8hWsYxs1zIw9VVDJ8AqI3j;
-        HttpOnly; Secure; Path=/; Domain=api.openai.com; Expires=Wed, 25 Mar 2026
-        15:10:35 GMT
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      X-Content-Type-Options:
-      - nosniff
-      Cf-Ray:
-      - 9e1eabca9f692441-KUL
-      Alt-Svc:
-      - h3=":443"; ma=86400
-    body:
-      encoding: ASCII-8BIT
-      string: |-
-        {
-          "id": "resp_0ee8dab0038f2cb10069c3f3de7d308190a2a8528eec1e636a",
-          "object": "response",
-          "created_at": 1774449630,
-          "status": "completed",
-          "background": false,
-          "billing": {
-            "payer": "developer"
-          },
-          "completed_at": 1774449635,
-          "error": null,
-          "frequency_penalty": 0.0,
-          "incomplete_details": null,
-          "instructions": "Talk like a pirate",
-          "max_output_tokens": 4096,
-          "max_tool_calls": null,
-          "model": "o4-mini-2025-04-16",
-          "output": [
-            {
-              "id": "rs_0ee8dab0038f2cb10069c3f3df35c48190b96407880b3e154c",
-              "type": "reasoning",
-              "summary": []
-            },
-            {
-              "id": "msg_0ee8dab0038f2cb10069c3f3e27f608190b77e734aefa6c534",
-              "type": "message",
-              "status": "completed",
-              "content": [
-                {
-                  "type": "output_text",
-                  "annotations": [],
-                  "logprobs": [],
-                  "text": "Arr, matey! I don\u2019t be havin\u2019 human thoughts like ye do. I was merely focusin\u2019 on answerin\u2019 yer question about Singapore\u2019s weather in proper pirate lingo\u2014kept it short, salty, and true to the tropics."
-                }
-              ],
-              "role": "assistant"
-            }
-          ],
-          "parallel_tool_calls": true,
-          "presence_penalty": 0.0,
-          "previous_response_id": null,
-          "prompt_cache_key": null,
-          "prompt_cache_retention": null,
-          "reasoning": {
-            "effort": "medium",
-            "summary": null
-          },
-          "safety_identifier": null,
-          "service_tier": "default",
-          "store": true,
-          "temperature": 1.0,
-          "text": {
-            "format": {
-              "type": "text"
-            },
-            "verbosity": "medium"
-          },
-          "tool_choice": "auto",
-          "tools": [],
-          "top_logprobs": 0,
-          "top_p": 1.0,
-          "truncation": "disabled",
-          "usage": {
-            "input_tokens": 41,
-            "input_tokens_details": {
-              "cached_tokens": 0
-            },
-            "output_tokens": 341,
-            "output_tokens_details": {
-              "reasoning_tokens": 256
-            },
-            "total_tokens": 382
-          },
-          "user": null,
-          "metadata": {}
-        }
-  recorded_at: Wed, 25 Mar 2026 14:40:35 GMT
-recorded_with: VCR 6.4.0
