--- test/fixtures/vcr_cassettes/integration/gateway_test_responses_test_rb/test_openai_responses_simple_message_without_tools.yml
@@ -1,155 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://api.openai.com/v1/responses
-    body:
-      encoding: UTF-8
-      string: '{"model":"o4-mini","max_output_tokens":4096,"input":[{"role":"user","content":[{"type":"input_text","text":"What''s
-        the weather in Singapore? reply in 10 words and no special characters"}]}],"instructions":"Talk
-        like a pirate"}'
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-      Host:
-      - api.openai.com
-      Content-Type:
-      - application/json
-      Authorization:
-      - Bearer <BEARER_TOKEN>
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Sat, 14 Feb 2026 14:45:07 GMT
-      Content-Type:
-      - application/json
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Server:
-      - cloudflare
-      X-Ratelimit-Limit-Requests:
-      - '500'
-      X-Ratelimit-Limit-Tokens:
-      - '200000'
-      X-Ratelimit-Remaining-Requests:
-      - '499'
-      X-Ratelimit-Remaining-Tokens:
-      - '199764'
-      X-Ratelimit-Reset-Requests:
-      - 120ms
-      X-Ratelimit-Reset-Tokens:
-      - 70ms
-      Openai-Version:
-      - '2020-10-01'
-      Openai-Organization:
-      - user-ayzusypxu6roh2akxcbajqo9
-      Openai-Project:
-      - proj_nmULvLHfn5yjMz2KqRbQcEQl
-      X-Request-Id:
-      - req_5d73794adff949a99fc2ca306887b56e
-      Openai-Processing-Ms:
-      - '6591'
-      Cf-Cache-Status:
-      - DYNAMIC
-      Set-Cookie:
-      - __cf_bm=hV3FGfm2HvST0B.hoqlwpjaXhiaKpch4_qkKaA09Lmg-1771080300.2062948-1.0.1.1-50HYhzSzGxqvu4OVa7IADmhKVJT_bU.uRhJHThAi7AxCRyrz2VnOfA_E49KGA0cmfT6wJwccarc2w0tnu1_9auGcqxTLgdrSX6Cts81Yccto5tXO3oNe2RqeEH7JcHy5;
-        HttpOnly; Secure; Path=/; Domain=api.openai.com; Expires=Sat, 14 Feb 2026
-        15:15:07 GMT
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      X-Content-Type-Options:
-      - nosniff
-      Cf-Ray:
-      - 9cdd58c44e95fd1e-SIN
-      Alt-Svc:
-      - h3=":443"; ma=86400
-    body:
-      encoding: ASCII-8BIT
-      string: |-
-        {
-          "id": "resp_0fcffb53adc177d90069908a6cbee4819199c62df7798425cb",
-          "object": "response",
-          "created_at": 1771080300,
-          "status": "completed",
-          "background": false,
-          "billing": {
-            "payer": "developer"
-          },
-          "completed_at": 1771080307,
-          "error": null,
-          "frequency_penalty": 0.0,
-          "incomplete_details": null,
-          "instructions": "Talk like a pirate",
-          "max_output_tokens": 4096,
-          "max_tool_calls": null,
-          "model": "o4-mini-2025-04-16",
-          "output": [
-            {
-              "id": "rs_0fcffb53adc177d90069908a6da5fc8191baf1db98833187f5",
-              "type": "reasoning",
-              "summary": []
-            },
-            {
-              "id": "msg_0fcffb53adc177d90069908a73033481918e219bdbd157c77b",
-              "type": "message",
-              "status": "completed",
-              "content": [
-                {
-                  "type": "output_text",
-                  "annotations": [],
-                  "logprobs": [],
-                  "text": "Arr me hearties weather be hot humid with brief showers"
-                }
-              ],
-              "role": "assistant"
-            }
-          ],
-          "parallel_tool_calls": true,
-          "presence_penalty": 0.0,
-          "previous_response_id": null,
-          "prompt_cache_key": null,
-          "prompt_cache_retention": null,
-          "reasoning": {
-            "effort": "medium",
-            "summary": null
-          },
-          "safety_identifier": null,
-          "service_tier": "default",
-          "store": true,
-          "temperature": 1.0,
-          "text": {
-            "format": {
-              "type": "text"
-            },
-            "verbosity": "medium"
-          },
-          "tool_choice": "auto",
-          "tools": [],
-          "top_logprobs": 0,
-          "top_p": 1.0,
-          "truncation": "disabled",
-          "usage": {
-            "input_tokens": 29,
-            "input_tokens_details": {
-              "cached_tokens": 0
-            },
-            "output_tokens": 732,
-            "output_tokens_details": {
-              "reasoning_tokens": 704
-            },
-            "total_tokens": 761
-          },
-          "user": null,
-          "metadata": {}
-        }
-  recorded_at: Sat, 14 Feb 2026 14:45:07 GMT
-recorded_with: VCR 6.4.0
