--- test/fixtures/vcr_cassettes/integration/gateway_test_responses_test_rb/test_openai_responses_weather_with_pirate_system.yml
@@ -1,171 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://api.openai.com/v1/responses
-    body:
-      encoding: UTF-8
-      string: '{"model":"o4-mini","max_output_tokens":4096,"input":[{"role":"user","content":[{"type":"input_text","text":"What''s
-        the weather in Singapore? reply in 10 words and no special characters"}]}],"instructions":"Talk
-        like a pirate","tools":[{"type":"function","name":"get_weather","description":"Get
-        current weather for a location","parameters":{"type":"object","properties":{"location":{"type":"string","description":"City
-        name"}},"required":["location"]}}]}'
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-      Host:
-      - api.openai.com
-      Content-Type:
-      - application/json
-      Authorization:
-      - Bearer <BEARER_TOKEN>
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Sat, 14 Feb 2026 14:45:14 GMT
-      Content-Type:
-      - application/json
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Server:
-      - cloudflare
-      X-Ratelimit-Limit-Requests:
-      - '500'
-      X-Ratelimit-Limit-Tokens:
-      - '200000'
-      X-Ratelimit-Remaining-Requests:
-      - '499'
-      X-Ratelimit-Remaining-Tokens:
-      - '199670'
-      X-Ratelimit-Reset-Requests:
-      - 120ms
-      X-Ratelimit-Reset-Tokens:
-      - 99ms
-      Openai-Version:
-      - '2020-10-01'
-      Openai-Organization:
-      - user-ayzusypxu6roh2akxcbajqo9
-      Openai-Project:
-      - proj_nmULvLHfn5yjMz2KqRbQcEQl
-      X-Request-Id:
-      - req_59cbff80e0d143529489b35c02391b7f
-      Openai-Processing-Ms:
-      - '6447'
-      Cf-Cache-Status:
-      - DYNAMIC
-      Set-Cookie:
-      - __cf_bm=T7r_wFepH9.TX8qvL.ds9Krniq0kcMVoFoXJvSLNhtQ-1771080308.3008409-1.0.1.1-AL2nyxa6itz2kY7bMFIFfKiSeOGvas6sclH1cnLwJInKGhyzIsIzbBh0H8qXlftsa6.6XMoa083gfwSypyK5B9m76FnSLKb_r2ukSOxZiy8XwxGim1A_Cxr1xes0B0qU;
-        HttpOnly; Secure; Path=/; Domain=api.openai.com; Expires=Sat, 14 Feb 2026
-        15:15:14 GMT
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      X-Content-Type-Options:
-      - nosniff
-      Cf-Ray:
-      - 9cdd58f6df04fd81-SIN
-      Alt-Svc:
-      - h3=":443"; ma=86400
-    body:
-      encoding: ASCII-8BIT
-      string: |-
-        {
-          "id": "resp_0ee686a33d6e715c0069908a746c708190a09722ba7695d606",
-          "object": "response",
-          "created_at": 1771080308,
-          "status": "completed",
-          "background": false,
-          "billing": {
-            "payer": "developer"
-          },
-          "completed_at": 1771080314,
-          "error": null,
-          "frequency_penalty": 0.0,
-          "incomplete_details": null,
-          "instructions": "Talk like a pirate",
-          "max_output_tokens": 4096,
-          "max_tool_calls": null,
-          "model": "o4-mini-2025-04-16",
-          "output": [
-            {
-              "id": "rs_0ee686a33d6e715c0069908a74c5f08190a5ec0dc5d639ad36",
-              "type": "reasoning",
-              "summary": []
-            },
-            {
-              "id": "fc_0ee686a33d6e715c0069908a7a8d088190950d769ce4c5653d",
-              "type": "function_call",
-              "status": "completed",
-              "arguments": "{\"location\":\"Singapore\"}",
-              "call_id": "call_LuLIijIGpw08QOK6vcEEBoHZ",
-              "name": "get_weather"
-            }
-          ],
-          "parallel_tool_calls": true,
-          "presence_penalty": 0.0,
-          "previous_response_id": null,
-          "prompt_cache_key": null,
-          "prompt_cache_retention": null,
-          "reasoning": {
-            "effort": "medium",
-            "summary": null
-          },
-          "safety_identifier": null,
-          "service_tier": "default",
-          "store": true,
-          "temperature": 1.0,
-          "text": {
-            "format": {
-              "type": "text"
-            },
-            "verbosity": "medium"
-          },
-          "tool_choice": "auto",
-          "tools": [
-            {
-              "type": "function",
-              "description": "Get current weather for a location",
-              "name": "get_weather",
-              "parameters": {
-                "type": "object",
-                "properties": {
-                  "location": {
-                    "type": "string",
-                    "description": "City name"
-                  }
-                },
-                "required": [
-                  "location"
-                ],
-                "additionalProperties": false
-              },
-              "strict": true
-            }
-          ],
-          "top_logprobs": 0,
-          "top_p": 1.0,
-          "truncation": "disabled",
-          "usage": {
-            "input_tokens": 67,
-            "input_tokens_details": {
-              "cached_tokens": 0
-            },
-            "output_tokens": 1015,
-            "output_tokens_details": {
-              "reasoning_tokens": 960
-            },
-            "total_tokens": 1082
-          },
-          "user": null,
-          "metadata": {}
-        }
-  recorded_at: Sat, 14 Feb 2026 14:45:15 GMT
-recorded_with: VCR 6.4.0
